Extracted MainViewModel to abstract class
